--- concordia/turnstile/context_processor.py
@@ -1,5 +1,0 @@
-from django.conf import settings
-
-
-def turnstile_settings(request):
-    return {"TURN_JS_API_URL": settings.TURN_JS_API_URL}
